requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
SQLAlchemy
aiohttp==3.9.1
asyncio-throttle==1.0.2
psutil==5.9.5
requests==2.31.0
numpy==1.24.3